영상 조회 관련 API

+++ business/src/main/java/com/mzbr/business/auth/dto/SignUpDto.java
@@ -12,10 +12,10 @@ import lombok.NoArgsConstructor;
 @Builder
 public class SignUpDto {
 
-	private final int userId;
+	private final long userId;
 	private final String nickname;
 
-	public static SignUpDto of(int userId, String nickname) {
+	public static SignUpDto of(long userId, String nickname) {
 		return SignUpDto.builder()
 			.userId(userId)
 			.nickname(nickname)
